Replace dead pixel viewer website
The "CameraCapture" and "CameraTestPattern" examples output raw image data for a frame captured from the camera. The
user is instructed to use the output with a website that provides a raw image display tool.

The website previously referenced went dead two years ago and the domain has been taken over by an unrelated entity. The
URL now redirects to a potentially malicious website.

The examples are hereby updated to suggest an alternative site. Although less full feature than the previous site, the
lost functionality was not relevant to our use case. Best of all, the new site is open source!